chore(demo): further refactoring
