+++ temporary/interfacing/JourneyGenerator.ts
@@ -1,9 +1,17 @@
-import AbstractGenerator from './AbstractGenerator';
+import Generator from './Generator';
 import PageGenerator from './PageGenerator';
-import { JSTemplate, JSMethodTemplate, TSTemplate, TSMethodTemplate, JSImportTemplate, TSImportTemplate, TSPageConstantTemplate } from './JourneyTemplates';
 
-export default class JourneyGenerator extends AbstractGenerator {
-    private _appPrefix: string;
+export default abstract class JourneyGenerator extends Generator {
+    _appPrefix: string;
+    _testName = "";
+    _steps: {
+        "step-comment": string,
+        "step-type": string,
+        "page-name": string,
+        "function-name": string,
+        "step"?: Record<string, unknown>
+    }[] = [];
+    _pages: Record<string, PageGenerator> = {};
 
     setJourneyJSON(oJourneyJSON: Record<string, unknown>): JourneyGenerator {
         this._extractAppPrefix(oJourneyJSON);
@@ -11,6 +19,42 @@ export default class JourneyGenerator extends AbstractGenerator {
         this._extractSteps(oJourneyJSON);
         return this;
     }
+
+    generate(bTS: boolean = false): string {
+        const sJourneyTemplate = this._getJourneyTemplate(bTS);
+        const sMethodTemplate = this._getMethodTemplate(bTS);
+
+        const placeholders = {
+            "journey-name": this._testName,
+            "test-intention": this._testName,
+            "app-prefix": this._appPrefix,
+            "page-first-name": Object.keys(this._pages).length > 0 ? Object.keys(this._pages)[0] : '<empty>',
+            "page-user": '\n' + Object.keys(this._pages).map(sP => this._replacePlaceholders(`const onThe{{page-name}}Page = new {{page-name}}Page();`.slice(), { "page-name": sP })).join("\n"),
+            "page-user-first": Object.keys(this._pages).length > 0 ? Object.keys(this._pages)[0] : '<empty>',
+            "page-import": this._getPageImports(bTS),
+            "step-insert": this._steps.map(oStep => {
+                const stepClone = { ...oStep };
+                delete stepClone.step;
+                return this._replacePlaceholders(sMethodTemplate.slice(), (stepClone as unknown as Record<string, string>));
+            }).join('\n')
+        }
+
+        return this._replacePlaceholders(sJourneyTemplate, placeholders);
+    }
+
+    generatePages(bTypeScript: boolean = false): { pageName: string, pageContent: string }[] {
+        return Object.entries(this._pages).map(eP => ({ pageName: eP[0], pageContent: (eP[1] as PageGenerator).generate(bTypeScript) }));
+    }
+
+    abstract _getPageGenerator(sPageName: string, sPageHash: string): PageGenerator;
+
+    abstract _getJourneyTemplate(bTS: boolean): string;
+
+    abstract _getMethodTemplate(bTS: boolean): string;
+
+    abstract _getImportTemplate(bTS: boolean): string;
+
+    abstract _getPageImports(bTS: boolean): string;
 
     private _extractAppPrefix(oJourneyJSON: Record<string, unknown>): JourneyGenerator {
         const aSteps = oJourneyJSON["steps"] as Record<string, unknown>[];
@@ -21,33 +65,35 @@ export default class JourneyGenerator extends AbstractGenerator {
         return this;
     }
 
-    generate(bTS: boolean = false): string {
-        const sJourneyTemplate = bTS ? TSTemplate : JSTemplate;
-        const sMethodTemplate = bTS ? TSMethodTemplate : JSMethodTemplate;
-        const sImportTemplate = bTS ? TSImportTemplate : JSImportTemplate;
-
-        const placeholders = {
-            "journey-name": this._testName,
-            "test-intention": this._testName,
-            "app-prefix": this._appPrefix,
-            "page-user": '\n' + Object.keys(this._pages).map(sP => this._replacePlaceholders(TSPageConstantTemplate.slice(), { "page-name": sP })).join("\n"),
-            "page-user-first": Object.keys(this._pages).length > 0 ? Object.keys(this._pages)[0] : '<empty>',
-            "page-import": (bTS ? "\n" : ",\n") + Object.keys(this._pages).map(sP => this._replacePlaceholders(sImportTemplate.slice(), { "page-name": sP })).join(bTS ? "\n" : ",\n"),
-            "step-insert": '\n' + this._steps.map(oStep => {
-                const stepClone = { ...oStep };
-                delete stepClone.step;
-                return this._replacePlaceholders(sMethodTemplate.slice(), (stepClone as unknown as Record<string, string>));
-            }).join('\n\n') + '\n'
-        }
-
-        return this._replacePlaceholders(sJourneyTemplate, placeholders);
+    private _extractJourneyName(oJourneyJSON: Record<string, unknown>): void {
+        this._testName = oJourneyJSON["name"] as string;
     }
 
-    generatePages(bTypeScript: boolean = false): { pageName: string, pageContent: string }[] {
-        return Object.entries(this._pages).map(eP => ({ pageName: eP[0], pageContent: (eP[1] as PageGenerator).generate(bTypeScript) }));
-    }
+    private _extractSteps(oJourneyJSON: Record<string, unknown>): void {
+        const aSteps = oJourneyJSON["steps"] as Record<string, unknown>[];
+        this._steps = aSteps.map((oStep: Record<string, unknown>) => {
+            const oViewInfos = oStep["viewInfos"] as { absoluteViewName: string, relativeViewName: string };
+            const bAssertion = oStep["actionType"] === 'validate';
+            const sMethodName = this._genMethodNameForStep(oStep);
+            const sComment = (bAssertion ? ' Assertion' : ' Action') + (oStep.comment ? `: ${oStep.comment}` : '');
+            const sType = bAssertion ? 'Then' : 'When';
+            const sPageName = oViewInfos.relativeViewName;
 
-    _getPageGenerator(sPageName: string, sPageHash: string): PageGenerator {
-        return new PageGenerator(sPageName, sPageHash);
+            if (!this._pages[sPageName]) {
+                const oViewInfos = oStep["viewInfos"] as { absoluteViewName: string, relativeViewName: string };
+                this._pages[sPageName] = this._getPageGenerator(oViewInfos.absoluteViewName, oStep.actionLocation as string);
+            }
+
+            (this._pages[sPageName] as PageGenerator).addMethod((oStep as Record<string, unknown>));
+
+            return {
+                "step-comment": sComment,
+                "step-type": sType,
+                //"step-selector": sStepSelector,
+                "page-name": sPageName,
+                "function-name": sMethodName,
+                "step": oStep
+            }
+        });
     }
 }
